Using doubles instead

+++ lab4/coordinate.h
@@ -2,17 +2,17 @@
 #define _coordinate_h
 
 struct cord {
-    float x0 ;
-    float x1 ;
-    float y0 ;
-    float y1 ;
+    double x0 ;
+    double x1 ;
+    double y0 ;
+    double y1 ;
 } ;
 
 struct part_cord {
-    float x ;
-    float y ;
-    float vx ;
-    float vy ;
+    double x ;
+    double y ;
+    double vx ;
+    double vy ;
 } ;
 
 typedef struct cord cord_t ;
