+++ lab4/physics.c
@@ -12,16 +12,16 @@
 #define sign(a) ((a) > 0 ? 1 : -1)
 #endif
 
-int feuler(pcord_t *a, float time_step)
+int feuler(pcord_t *a, double time_step)
 {
     a->x = a->x + time_step * a->vx ;
     a->y = a->y + time_step * a->vy ;
     return 0 ;
 }
 
-float wall_collide(pcord_t *p, cord_t wall)
+double wall_collide(pcord_t *p, cord_t wall)
 {
-    float gPreassure = 0.0 ;
+    double gPreassure = 0.0 ;
 
     if (p->x < wall.x0)
     {
@@ -52,7 +52,7 @@ float wall_collide(pcord_t *p, cord_t wall)
 
 
 
-float collide(pcord_t *p1, pcord_t *p2)
+double collide(pcord_t *p1, pcord_t *p2)
 {
     double a, b, c;
     double temp, t1, t2;
@@ -87,9 +87,9 @@ float collide(pcord_t *p1, pcord_t *p2)
 
 
 
-void interact(pcord_t *p1, pcord_t *p2, float t)
+void interact(pcord_t *p1, pcord_t *p2, double t)
 {
-    float c, s, a, b, tao;
+    double c, s, a, b, tao;
     pcord_t p1temp, p2temp;
 
     if (t >= 0)
